Remove deps.edn and lein-deps-tools.
Using lein-deps-tools poses
a challenge similar to the one described here
https://github.com/RickMoynihan/lein-tools-deps/issues/79,
but with Github actions.

Owing to this change, going foward, the dependency closest to the root of the
dependency tree will be used (which is the default behaviour
with leiningen) instead of the latest dependency when there are conflicts between
lein and maven's depedency resolution algortihm.